Update P1
Design and implement a computer player that can play the Isolation game independently - this is a double sure game that turns in the regular grid until one of them can not move on. Use the heuristic evaluation of the heuristic search to overcome the computational problems and make the computer players more intelligent.